Move history utility to cache module
